New disk builder
It allows for up to 30 files, which can be specified in buildlist.txt
and includelist.txt. Files in buildlist.txt assume a starting dir of
src/ and will be built before being included. Files in includelist.txt
will assume a starting dir of bin/ and will be directly included as
binaries. Please properly capitalize filenames and paths for
compatibility with linux. More documentation to follow.