Made build.py use diskdata.dasm
This means it no longer generates a FAT or DTEs for in the root dir.
You'll have to do all that yourself. It will still include all files
specified in buildlist.txt and includelist.txt, using the same order as
in those files, the ones in buildlist.txt first. The FAT and DTEs for
all directories need to be put into diskdata.dasm. You can calculate the
offset of files because you know how big they all are and in which order
they are. NOT TESTED.